# api/api.yml
openapi: "3.0.0"
info:
  title: "eyevow api"
  description: ""
  version: "1.0.0"
servers:
  - url: https://eyevow.com/api

# ----
# エンドポイント
# ----
paths:
...
  /api/vow:
    get:
      security:
        - AccessToken: []
      tags:
        - 誓い
      summary: 現在の誓いを確認する
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                properties:
                  vow:
                    $ref: '#/components/schemas/Vow'
    put:
      tags:
        - 誓い
      summary: 誓いを立てる
      requestBody:
        content:
          application/json:
            schema:
              properties:
                text:
                  type: string
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                properties:
                  vow:
                    $ref: '#/components/schemas/Vow'
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      properties:
        id:
          type: string
    Vow:
      properties:
        id:
          type: string
        user:
          $ref: '#/components/schemas/User'
        text:
          type: string
        cheer_count:
          type: integer
        archived:
          type: boolean